update: hero section
